draw shape text while dragging
